requirements.txt
google-genai>=1.0.0
arxiv>=2.1.0
pyyaml>=6.0
python-telegram-bot>=21.0
